% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{avonet}
\alias{avonet}
\title{AVONET trait list}
\format{
DataFrame
\describe{
  \item{species}{species name}
  \item{order}{taxonomic order}
  \item{wing_length}{wing length (in mm)}
  \item{secondary}{length of the secondary (in mm)}
  \item{mass}{body mass (in gr)}
}
}
\usage{
avonet
}
\description{
Limited copy of the AVONET trait database used
to calculate flight dynamic characteristics.
}
\keyword{datasets}
