% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bird_create.R
\name{bird_create}
\alias{bird_create}
\title{Create bird flight traits}
\usage{
bird_create(
  scientific_name,
  mass = NULL,
  wing_span = NULL,
  wing_aspect = NULL,
  wing_area = NULL,
  body_frontal_area = NULL
)
}
\arguments{
\item{scientific_name}{Scientific name of the species}

\item{mass}{Mass of the bird in kilograms}

\item{wing_span}{Wing span in meters}

\item{wing_aspect}{Wing aspect ratio (no unit)}

\item{wing_area}{Wing area in meter square}

\item{body_frontal_area}{Body frontal area in meter square}

\item{species_name}{`r lifecycle::badge("deprecated")` Use `scientific_name` instead}
}
\value{
List containing mass, wing span, wing aspect ratio, and body frontal area.
}
\description{
This function returns a list with the four morphological information necessary to construct the
power curve: mass, wing span, wing aspect ratio, and body frontal area.

When any of these variables are missing, we query the
[AVONET database](https://doi.org/10.6084/m9.figshare.16586228.v5) using the scientific
name from [the Clements Checklist](https://www.birds.cornell.edu/clementschecklist/).
}
\examples{
# Using AVONET dataset
bird_create("Acrocephalus arundinaceus")

# Using AVONET dataset + custom values
bird_create("Acrocephalus arundinaceus", wing_aspect = 8)

# Import your own bird. You will need mass, and at least two of wing_span,
# wing_aspect or wing_area.
bird_create("Madynuss nutshell", mass = 8, wing_span = 0.2, wing_aspect = 4)

}
\seealso{
Other movement: 
\code{\link{speed2prob}()}
}
\concept{bird}
\concept{movement}
