% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cc876}
\alias{cc876}
\title{Migrate Technology Ltd demo data}
\format{
DataFrame
\describe{
  \item{logger}{logger ID}
  \item{date}{date}
  \item{date_time}{date and time}
  \item{hour}{decimal hour}
  \item{lux}{light levels in lux}
}
}
\usage{
cc876
}
\description{
Demo data for a single day of Common swift light logger
data as read from a Migrate Technology Ltd .lux file
using stk_read_lux().
}
\details{
The format is consistent with what is required by
the skytrackr() routine.
}
\keyword{datasets}
