% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/illuminance_models.R
\name{diurnal}
\alias{diurnal}
\title{Simulate diurnal illuminance value}
\usage{
diurnal(par, data, loc, ...)
}
\arguments{
\item{par}{Three parameters specifying the illuminance model.}

\item{data}{A data frame with the required drivers for the illuminance model.}

\item{...}{optional other parameters to forward}
}
\value{
Sky illuminance as log(lux).
}
\description{
Calculates log(lux) values for a give location, date, time
and sky conditions.
}
