% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/illuminance_models.R
\name{individual}
\alias{individual}
\title{Updates values individually not globally (diurnally)}
\usage{
individual(par, data, loc, ...)
}
\arguments{
\item{par}{Three parameters specifying the illuminance model.}

\item{data}{A data frame with the required drivers for the illuminance model.}

\item{loc}{previous location}

\item{...}{optional other parameters to forward}
}
\value{
Sky illuminance as log(lux).
}
\description{
Calculates log(lux) values for a give location, date, time
and sky conditions along the track. Updates to the bearing
direction of flight are only made for selected values.
}
\details{
After a diurnal cycle the course is updated for the next,
diurnal cycle. Processing uses a GMT time reference and
large east-west movements might cut through the date line,
hence updates along the flight track would not perfectly
coincide with the end of a twilight phase.
}
