% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{land}
\alias{land}
\title{Land area vector}
\format{
sf
\describe{
  \item{MULTIPOLYGON}{sf multipolygon}
}
}
\usage{
land
}
\description{
Vector polygon of world land areas to constrain
model optimization.
}
\keyword{datasets}
