% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likelihood.R
\name{likelihood}
\alias{likelihood}
\title{Log likelihood cost function for model optimization}
\usage{
likelihood(par, data, model, loc, roi, step_selection, ...)
}
\arguments{
\item{par}{a vector of parameter values, including one for the uncertainty
on the target values}

\item{data}{nested data structure with validation data included}

\item{model}{model to run with data and par settings}

\item{loc}{previous modeled step location}

\item{roi}{region of interest with valid sampling locations}

\item{step_selection}{a step selection function on the distance of a proposed move}

\item{...}{extra arguments to pass to the function}
}
\value{
single log likelihood
}
\description{
Log likelihood cost function for model optimization
}
\keyword{cost}
\keyword{function}
\keyword{model,}
\keyword{optimization,}
