% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log_lux.R
\name{log_lux}
\alias{log_lux}
\title{Simulate illuminance value}
\usage{
log_lux(par, data, ...)
}
\arguments{
\item{par}{three parameters specifying the cycling model}

\item{data}{data consisting of vegetation greenness (G), mean
daytime temperature (T) and daylenght (D) as a data frame}

\item{...}{optional other parameters to forward}
}
\value{
sky illuminance as log(lux)
}
\description{
Calculates log(lux) values for a give location, date, time
and sky conditions.
}
