% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/movement_kernel.R
\name{movement_kernel}
\alias{movement_kernel}
\title{Convert roi mask to movement kernel}
\usage{
movement_kernel(roi, shape = 0.9, scale = 2)
}
\arguments{
\item{roi}{region of interest raster file}

\item{shape}{weibull shape parameter}

\item{scale}{weibull scale parameter}
}
\value{
weibull log likelihoods (approximation)
}
\description{
Applies a distance matrix and weibull
density function to a region of interest
}
\details{
This is approximative of probability density
data, good enough for now. Should be
corrected for cell spacing / count,
see Ranc et al. 2022 Ecology Letters
}
