% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stk_read.R
\name{read_deg_lux}
\alias{read_deg_lux}
\title{Read lux and deg files}
\usage{
read_deg_lux(file, verbose = TRUE)
}
\arguments{
\item{file}{A lux or deg file.}
}
\value{
A skytrackr data frame with logger data.
}
\description{
This function is wrapped by the `stk_read_lux()` function.
}
