% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{rivers}
\alias{rivers}
\title{River lines}
\format{
sf
\describe{
  \item{MULTILINESTRING}{sf multilinestring}
}
}
\usage{
rivers
}
\description{
Vector line strings of world rivers
}
\keyword{datasets}
