% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skytrack.R
\name{skytrack}
\alias{skytrack}
\title{Sky (illuminance) tracker}
\usage{
skytrack(
  data,
  iterations = 10000,
  floor = 0.32,
  ceiling = 400,
  bbox = c(-180, -90, 180, 90),
  offset = "day",
  plot = FALSE
)
}
\arguments{
\item{data}{a data frame containing date time and lux values}

\item{iterations}{number of optimization iterations}

\item{floor}{lower bound of lux values to consider (values below this
value are ignored)}

\item{ceiling}{upper bound of lux values to consider (values above this
value are ignored)}

\item{bbox}{bounding box of the location search domain given as
c(xmin, ymin, xmax, ymax)}

\item{offset}{by default the diurnal cycle around noon is considered (from
sunrise to sunset), alternatively estimates can be made around  midnight
considering the profile from sunset to sunrise.}

\item{plot}{plot map of incrementally changing determined locations as
a progress method}
}
\value{
data frame with location estimate, their uncertainties, and
ancillary model parameters useful in quality control
}
\description{
Skytrack compares geolocator based light measurements in lux with
those modelled by the sky illuminance model of xyz.
}
\details{
Model fits are applied by default to values up to sunrise or after
sunset only as most critical to the model fit (capturing daylength,
i.e. latitude and the location of the diurnal pattern -
longitudinal displacement).
}
