% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step_selection_probability.R
\name{speed2prob}
\alias{speed2prob}
\title{Compute probability of a bird speed}
\usage{
speed2prob(speed, movement)
}
\arguments{
\item{speed}{airspeed or ground speed in km/h}

\item{movement}{a list of the movement model parameter defined with `graph_set_movement`}
}
\value{
Probability values corresponding to the speed provided.
}
\description{
Converts a speed (airspeed or ground speed) to a
probability based on the movement model.
}
\examples{
speed <- seq(1, 120)
low_speed_fix <- 20 # minimum speed allowed
prob <- speed2prob(
  speed,
  list(
    method = "gamma",
    shape = 7,
    scale = 7,
    low_speed_fix = low_speed_fix
  )
)
plot(speed, prob,
  type = "l",
  xlab = "Groundspeed [km/h]",
  ylab = "Probability"
)
abline(v = low_speed_fix)

# Using airspeed
bird <- bird_create("Acrocephalus arundinaceus")
prob <- speed2prob(
  speed,
  list(
    method = "power",
    bird = bird,
    power2prob = \(power) (1 / power)^3,
    low_speed_fix = low_speed_fix
  )
)
plot(speed, prob, type = "l", xlab = "Airspeed [km/h]", ylab = "Probability")
abline(v = low_speed_fix)
}
\seealso{
Other movement: 
\code{\link{bird_create}()}
}
\concept{movement}
