% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/movement_kernel.R
\name{step_selection}
\alias{step_selection}
\title{Step selection}
\usage{
step_selection(par, ref, shape = 1.02, scale = 150000)
}
\arguments{
\item{par}{(random) parameter location coordinates}

\item{ref}{reference location coordinates}

\item{shape}{gamma distribution shape parameter}

\item{scale}{gamma distribution scale parameter}
}
\value{
log-likelihood of a step distance given a gamma distribution
}
\description{
Step selection probability density function
based upon the distance between a reference
point (previous step) and a new parameter location.
}
\details{
Step selection will follow a gamma distribution to
be defined with scale and shape parameters (either
through fitting known data, e.g. GPS data, or by
making reasonable assumption on behaviour).
}
