% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stk_center.R
\name{stk_center}
\alias{stk_center}
\title{Centers diurnal curves on midday}
\usage{
stk_center(df, floor = 1.5, replace = FALSE)
}
\arguments{
\item{df}{a skytrackr compatible data frame}

\item{floor}{threshold value to center the data with (default = 1.5)}

\item{replace}{replace original decimal hour values (default = FALSE)}
}
\value{
a day-by-day centered skytrackr compatible data frame
}
\description{
Daily values are centered on midday on a day-by-day basis. Note that this does
not shift the time series properly as hours are wrapped around on a given day.
This function should not be used stand-alone, and is primarily used to screen
data for noise / poor quality.
}
\examples{

cc876 |> stk_center()
}
