% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stk_profile.R
\name{stk_profile}
\alias{stk_profile}
\title{Plot seasonal profiles}
\usage{
stk_profile(data, logger, plotly = FALSE)
}
\arguments{
\item{data}{A skytrackr compatible data frame.}

\item{logger}{The logger to plot}

\item{plotly}{Logical, convert to dynamic plotly plot or not (default = FALSE)}
}
\value{
A static or dynamic graph of light levels for a given logger.
}
\description{
Provides static or dynamic (plotly) seasonal profile plot
}
