% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stk_read.R
\name{stk_read_glf}
\alias{stk_read_glf}
\title{Read GLF files}
\usage{
stk_read_glf(file)
}
\arguments{
\item{file}{a .glf file with light level values}
}
\value{
a skytrackr compatible data frame for use in further location
 estimation
}
\description{
Read Swiss Ornithology institute files
}
