% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stk_read_lux.R
\name{stk_read_lux}
\alias{stk_read_lux}
\title{Read Migrate Technology .lux files}
\usage{
stk_read_lux(file)
}
\arguments{
\item{file}{a .lux file with light level values}
}
\value{
a skytrackr compatible data frame for use in further location
 estimation
}
\description{
Read Migrate Technology lux files and re-formats
them to a skytrackr compatible format.
}
