% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stk_screen_twl.R
\name{stk_screen_twl}
\alias{stk_screen_twl}
\title{Twilight screening routine}
\usage{
stk_screen_twl(df, threshold = 1.5, dips = 3, step = 100, filter = FALSE)
}
\arguments{
\item{df}{skytrackr data frame}

\item{threshold}{twilight threshold (default = 1.5)}

\item{dips}{allowed number of interruptions during a daylight profile
below the twilight threshold before flagging as a poor quality "suspect"
day}

\item{step}{threshold of the allowed step change in illuminance values
between the twilight value and the preceding one. Large jumps and the lack
of a smooth transition suggest a false twilight (bird leaving a dark nest
site long after or long before dawn or dusk).}

\item{filter}{return the data prefiltered, removing all poor quality days
or false twilight ones}
}
\value{
a skytrackr data frame with poor twilight quality days removed and
 dusk and dawn timings marked (data is returned as a long format, not a wide
 format)
}
\description{
Removes poor quality data based on twilight heuristics
}
