% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likelihood.R
\name{likelihood}
\alias{likelihood}
\title{Log likelihood cost function for model optimization}
\usage{
likelihood(par, data, model, mask, ...)
}
\arguments{
\item{par}{a vector of parameter values, including one for the uncertainty
on the target values}

\item{data}{nested data structure with validation data included}

\item{model}{model to run with data and par setings}

\item{mask}{mask to use when running the model}

\item{...}{extra arguments to pass to the function}
}
\value{
single log likelihood
}
\description{
The function is aimed to be maximized, to use it with optimizers which
minimize cost functions wrap the function as such:
`cost = function(...){abs(likelihood(...))}`
}
\keyword{cost}
\keyword{function}
\keyword{model,}
\keyword{optimization,}
