% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likelihood.R
\name{likelihood}
\alias{likelihood}
\title{Log likelihood cost function}
\usage{
likelihood(par, data, model, loc, roi, step_selection, ...)
}
\arguments{
\item{par}{A vector of parameter values, including one for the uncertainty
on the target values.}

\item{data}{A nested data structure with validation data included.}

\item{model}{A model to run with data and par settings.}

\item{loc}{The previous modeled step location.}

\item{roi}{A region of interest with valid sampling locations.}

\item{step_selection}{A step selection function on the distance of a proposed move.}

\item{...}{extra arguments to pass to the function}
}
\value{
The single log-likelihood cost of a proposed parameter set.
}
\description{
Main cost function used during optimization, combining both the fit of the
illuminance data with the step-selection function.
}
