% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log_lux.R
\name{log_lux}
\alias{log_lux}
\title{Simulate illuminance value}
\usage{
log_lux(par, data, ...)
}
\arguments{
\item{par}{Three parameters specifying the illuminance model.}

\item{data}{A data frame with the required drivers for the illuminance model.}

\item{...}{optional other parameters to forward}
}
\value{
Sky illuminance as log(lux).
}
\description{
Calculates log(lux) values for a give location, date, time
and sky conditions.
}
