% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stk_read.R
\name{read_deg_lux}
\alias{read_deg_lux}
\title{Read lux and deg files}
\usage{
read_deg_lux(file)
}
\arguments{
\item{file}{a lux or deg file}
}
\value{
a data frame with logger data
}
\description{
Read lux and deg files
}
