% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skytrackr.R
\name{skytrackr}
\alias{skytrackr}
\title{Sky (illuminance) tracker}
\usage{
skytrackr(
  data,
  start_location,
  tolerance = 1500,
  range = c(0.09, 148),
  scale = log(c(1e-05, 50)),
  control = list(sampler = "DEzs", settings = list(burnin = 250, iterations = 3000,
    message = FALSE)),
  mask,
  step_selection,
  plot = TRUE,
  verbose = TRUE
)
}
\arguments{
\item{data}{a data frame containing date time and lux values}

\item{start_location}{start location of logging, required when using a
tolerance based estimation. When no start_location is used the default
bounding box (bbox) settings are used.}

\item{tolerance}{tolerance distance on the search window for optimization,
given in km (left/right, top/bottom). Sets a hard limit on the search window
regardless of the step selection function used.}

\item{range}{range of values to consider during processing, should be
provided in lux c(min, max) or the equivalent if non-calibrated. In case of
non calibrated values you will need to adapt the scale values accordingly.}

\item{scale}{scale / sky condition factor, by default covering the
skylight() range of 1-10 (from clear sky to extensive cloud coverage)
but can be extended for more flexibility to account for coverage by plumage,
note that in case of non-physical accurate lux measurements values can have
a range starting at 0.0001 (a multiplier instead of a divider).}

\item{control}{control settings for the Bayesian optimization, generally
should not be altered (defaults to a Monte Carlo method)}

\item{mask}{mask with priors to constrain positions}

\item{step_selection}{a step selection function on the distance of a proposed
move, step selection is specified as average flight speed to achieve this
distance (in km/h).}

\item{plot}{plot map of incrementally changing determined locations as
a progress method}

\item{verbose}{given feedback including a progress bar}
}
\value{
data frame with location estimate, their uncertainties, and
ancillary model parameters useful in quality control
}
\description{
Skytrack compares geolocator based light measurements in lux with
those modelled by the sky illuminance model of xyz.
}
\details{
Model fits are applied by default to values up to sunrise or after
sunset only as most critical to the model fit (capturing daylength,
i.e. latitude and the location of the diurnal pattern -
longitudinal displacement).
}
