% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skytrackr.R
\name{skytrackr}
\alias{skytrackr}
\title{Sky (illuminance) tracker}
\usage{
skytrackr(
  data,
  iterations = 20000,
  range = c(0.32, 400),
  bbox = c(-180, -90, 180, 90),
  scale = c(1, 10),
  offset = "day",
  control = list(sampler = "DEzs", settings = list(burnin = iterations * 0.2, iterations
    = iterations * 0.8, message = FALSE)),
  plot = FALSE,
  verbose = TRUE
)
}
\arguments{
\item{data}{a data frame containing date time and lux values}

\item{iterations}{number of optimization iterations}

\item{range}{range of values to consider during processing, should be
provided in lux c(min, max) or the equivalent if non-calibrated}

\item{bbox}{bounding box of the location search domain given as
c(xmin, ymin, xmax, ymax)}

\item{scale}{scale / sky condition factor, by default covering the
skylight() range of 1-10 but can be extended for more flexibility
in case of non lux measurements}

\item{offset}{by default the diurnal cycle around noon is considered (from
sunrise to sunset), alternatively estimates can be made around  midnight
considering the profile from sunset to sunrise.}

\item{control}{control settings for the Bayesian optimization, generally
should not be altered}

\item{plot}{plot map of incrementally changing determined locations as
a progress method}

\item{verbose}{given feedback including a progress bar}
}
\value{
data frame with location estimate, their uncertainties, and
ancillary model parameters useful in quality control
}
\description{
Skytrack compares geolocator based light measurements in lux with
those modelled by the sky illuminance model of xyz.
}
\details{
Model fits are applied by default to values up to sunrise or after
sunset only as most critical to the model fit (capturing daylength,
i.e. latitude and the location of the diurnal pattern -
longitudinal displacement).
}
