% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skytrackr.R
\name{skytrackr}
\alias{skytrackr}
\title{Sky (illuminance) tracker}
\usage{
skytrackr(
  data,
  start_location,
  tolerance = 15,
  iterations = 20,
  particles = 200,
  range = c(0.32, 400),
  bbox = c(-180, -90, 180, 90),
  scale = c(0, 20),
  control = list(sampler = "SMC", settings = list(initialParticles = 100, iterations =
    10, message = FALSE)),
  plot = FALSE,
  verbose = TRUE
)
}
\arguments{
\item{data}{a data frame containing date time and lux values}

\item{start_location}{start location of logging, required when using a
tolerance based estimation. When no start_location is used the default
bounding box (bbox) settings are used.}

\item{tolerance}{tolerance on the search window for optimization, given in
degrees (to pad left - right, top - bottom). Uses the previous steps
location to constrain the parameter (location) search space.
This requires the start_location parameter to be set, as you need a
start position for a trusted initial location.}

\item{iterations}{number of optimization iterations}

\item{particles}{number of particles when using the default particle filter
optimization}

\item{range}{range of values to consider during processing, should be
provided in lux c(min, max) or the equivalent if non-calibrated}

\item{bbox}{bounding box of the location search domain given as
c(xmin, ymin, xmax, ymax)}

\item{scale}{scale / sky condition factor, by default covering the
skylight() range of 1-10 but can be extended for more flexibility
in case of non lux measurements}

\item{control}{control settings for the Bayesian optimization, generally
should not be altered (defaults to a sequential monte carlo method)}

\item{plot}{plot map of incrementally changing determined locations as
a progress method}

\item{verbose}{given feedback including a progress bar}
}
\value{
data frame with location estimate, their uncertainties, and
ancillary model parameters useful in quality control
}
\description{
Skytrack compares geolocator based light measurements in lux with
those modelled by the sky illuminance model of xyz.
}
\details{
Model fits are applied by default to values up to sunrise or after
sunset only as most critical to the model fit (capturing daylength,
i.e. latitude and the location of the diurnal pattern -
longitudinal displacement).
}
