% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stk_calibrate.R
\name{stk_calibrate}
\alias{stk_calibrate}
\title{Estimate upper scale value}
\usage{
stk_calibrate(df, percentile = 100, floor = 1.5, verbose = TRUE)
}
\arguments{
\item{df}{a skytrackr dataframe}

\item{percentile}{percentile of the spread of data to use in scale value
evaluation}

\item{floor}{threshold to remove low (nighttime) values}

\item{verbose}{Give detailed feedback (TRUE or FALSE, default = TRUE)}
}
\value{
an estimated scale value to be used in optimization
}
\description{
Estimates the upper scale value based upon
midday values of the data itself. The method
requires midday values and will not work on
loggers with a capped value (only recording twilight
values)
}
\examples{

# Estimate the upper scale value for fitting routine
upper_scale_value <- cc876 |> stk_calibrate()
}
