% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stk_calibrate.R
\name{stk_calibrate}
\alias{stk_calibrate}
\title{Estimate scale the scale range}
\usage{
stk_calibrate(df, percentile = 100, floor = 1.5, verbose = TRUE)
}
\arguments{
\item{df}{a skytrackr dataframe}

\item{percentile}{percentile of the spread of data to use in scale value
evaluation}

\item{floor}{threshold to remove low (nighttime) values}

\item{verbose}{Give detailed feedback (TRUE or FALSE, default = TRUE)}
}
\value{
an estimated scale value to be used in optimization
}
\description{
Provides a rough estimate on the light loss and corresponding
range of scale values to consider during optimization. The full
range is reported, but a percentile parameter can be provided
to remove outliers. It is recommended to first use `st_screen_twl()`
to remove "odd" days.
}
\examples{

# Estimate the upper scale value for fitting routine
upper_scale_value <- cc876 |> stk_calibrate()
}
