% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stk_calibrate.R
\name{stk_calibrate}
\alias{stk_calibrate}
\title{Estimate upper scale value}
\usage{
stk_calibrate(df, floor = 1.5, sky_condition = 3, plot = TRUE, verbose = TRUE)
}
\arguments{
\item{df}{a skytrackr dataframe}

\item{floor}{threshold to remove low (nighttime) values}

\item{sky_condition}{reference sky conditions, by default set
to three (3), an average sky with ~30% cloud cover. One (1)
is a clear sky, while a value of ten (10) is an extremely dark
sky under stormy / rainy conditions (uncommon).}

\item{plot}{Plot statistical information on the suggested scale estimate}

\item{verbose}{Give feedback including a progress bar (TRUE or FALSE,
default = TRUE)}
}
\value{
an estimated scale value to be used in optimization
}
\description{
Estimates the upper scale value based upon
midday values of the data itself. The method
requires midday values and will not work on
loggers with a capped value (only recording twilight
values)
}
