% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stk_cluster.R
\name{stk_cluster}
\alias{stk_cluster}
\title{Cluster geolocator light levels}
\usage{
stk_cluster(df, eps = 0.1, plot = TRUE)
}
\arguments{
\item{df}{A skytrackr data frame}

\item{eps}{dbscan eps value, see `dbscan::dbscan()` (default = 0.1, condensed clusters)}

\item{plot}{plot the cluster results with respect to the diurnal pattern offset and (day)length values}
}
\value{
The original data frame with attached cluster labels.
}
\description{
Uses k-means to group geolocator light values
into consistent groups for further analysis
}
