% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stk_cluster.R
\name{stk_cluster}
\alias{stk_cluster}
\title{Cluster geolocator data}
\usage{
stk_cluster(df, k, center = TRUE)
}
\arguments{
\item{df}{a skytrackr data frame}

\item{k}{number of k-means clusters to consider}

\item{center}{center values or use raw numbers}
}
\value{
original data frame with attached cluster labels
}
\description{
Uses k-means clustering to group geolocator
covariates into consistent groups
}
