% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stk_filter.R
\name{stk_filter}
\alias{stk_filter}
\title{Filter twilight values by range}
\usage{
stk_filter(data, range, plot = FALSE, filter = FALSE)
}
\arguments{
\item{data}{a skytrackr compatible data frame}

\item{range}{a range c(min, max) of valid values in lux}

\item{plot}{plot daily profiles with the range filter applied}

\item{filter}{if TRUE only twilight values are returned if
FALSE the data frame is returned with an annotation column
called 'twilight' for further processing.}
}
\value{
a skytrackr compatible data frame, either filtered
 to only include twilight values selected by the range parameter
 or with an additional 'twilight' column to annotate these values.
}
\description{
Filter out twilight values by range, and
returns the data frame with a twilight (TRUE/FALSE)
column or the data frame with only twilight values
selected (filtered out).
}
\details{
Generally used for internal process, but can be
useful for visualizations of profiles as well.
}
