% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stk_fit.R
\name{stk_fit}
\alias{stk_fit}
\title{Fit illuminance (lux) profile}
\usage{
stk_fit(
  data,
  iterations,
  bbox = c(-180, -90, 180, 90),
  scale = c(1, 10),
  control = list(sampler = "DEzs", settings = list(burnin = iterations * 0.2, iterations
    = iterations * 0.8, message = FALSE))
)
}
\arguments{
\item{data}{a data frame containing date time and lux values}

\item{iterations}{number of optimization iterations}

\item{bbox}{bounding box of the location search domain given as
c(xmin, ymin, xmax, ymax)}

\item{scale}{scale factor range due to cloudiness to use in optimization,
when target values are not provided in lux this can be used to effectively
implement a Hill-Ekstrom template fitting approach}

\item{control}{control settings for the Bayesian optimization}
}
\value{
an estimated illuminance based location (and its uncertainties)
}
\description{
Fits a simulated lux profile to observed light logger data
to estimate locations (parameters).
}
