% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stk_fit.R
\name{stk_fit}
\alias{stk_fit}
\title{Fit illuminance (lux) profile}
\usage{
stk_fit(data, roi, loc, scale, control, step_selection)
}
\arguments{
\item{data}{a data frame containing date time and lux values}

\item{roi}{region of interest}

\item{loc}{location of the previous step}

\item{scale}{scale factor range due to cloudiness to use in optimization,
when target values are not provided in lux this can be used to effectively
implement a Hill-Ekstrom template fitting}

\item{control}{control settings for the Bayesian optimization, forwarded by
skytrackr()}

\item{step_selection}{a step selection function on the distance of a proposed move}
}
\value{
an estimated illuminance based location (and its uncertainties)
}
\description{
Fits a simulated lux profile to observed light logger data
to estimate locations (parameters).
}
