% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stk_map.R
\name{stk_map}
\alias{stk_map}
\title{Plot skytrackr results}
\usage{
stk_map(df, bbox, start_location, roi, dynamic = FALSE)
}
\arguments{
\item{df}{a data frame with locations produced by skytrackr()}

\item{bbox}{a bounding box}

\item{start_location}{start location as lat/lon to indicate
the starting position of the track (optional)}

\item{roi}{region of interest under consideration, only used in
plots during optimization (optional)}

\item{dynamic}{create a dynamic interactive graph rather than
a static plot. Both the path as the locations are shown. The size
of the points is proportional to the latitudinal uncertainty, while
equinox windows are marked with red points.}
}
\value{
a ggplot map of tracked locations
}
\description{
Create a map of estimated locations as a static or dynamic map.
}
\examples{
\dontrun{

#----- estimated locations on demo data ----
# define land mask with a bounding box
# and an off-shore buffer (in km), in addition
# you can specify the resolution of the resulting raster
mask <- stk_mask(
  bbox  =  c(-20, -40, 60, 60), #xmin, ymin, xmax, ymax
  buffer = 150, # in km
  resolution = 0.5 # map grid in degrees
  )

  # define a step selection distribution/function
  ssf <- function(x, shape = 0.9, scale = 100, tolerance = 1500){
  norm <- sum(stats::dgamma(1:tolerance, shape = shape, scale = scale))
  prob <- stats::dgamma(x, shape = shape, scale = scale) / norm
}

# estimate locations with default values
# and plot progress
locations <- cc876 |> skytrackr(plot = FALSE)

#----- actual plotting routines ----
# static plot, with required bounding box
locations |> stk_map(bbox = c(-20, -40, 60, 60))

# dynamic plot
locations |> stk_map(dynamic = TRUE)
}
}
