% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stk_map.R
\name{stk_map}
\alias{stk_map}
\title{Plot skytrackr results}
\usage{
stk_map(df, buffer, bbox, start_location, roi)
}
\arguments{
\item{df}{a data frame with locations produced by skytrackr()}

\item{buffer}{a land mask buffer value}

\item{bbox}{a bounding box}
}
\value{
a ggplot map of tracked locations
}
\description{
Create a map of estimated locations
}
