% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stk_map.R
\name{stk_map}
\alias{stk_map}
\title{Plot skytrackr results}
\usage{
stk_map(df, bbox, start_location, roi, dynamic = FALSE)
}
\arguments{
\item{df}{a data frame with locations produced by skytrackr()}

\item{bbox}{a bounding box}

\item{start_location}{start location as lat/lon to indicate
the starting position of the track (optional)}

\item{roi}{region of interest under consideration, only used in
plots during optimization (optional)}

\item{dynamic}{create a dynamic interactive graph rather than
a static plot. Both the path as the locations are shown. The size
of the points is proportional to the latitudinal uncertainty, while
equinox windows are marked with red points.}
}
\value{
a ggplot map of tracked locations
}
\description{
Create a map of estimated locations
}
