% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stk_mask.R
\name{stk_mask}
\alias{stk_mask}
\title{A land mask}
\usage{
stk_mask(buffer = 0, resolution = 1, bbox, sf = FALSE)
}
\arguments{
\item{buffer}{buffer in degrees}

\item{resolution}{resolution of the spatial grid in degrees, when exporting
as a terra SpatRaster (default = 1)}

\item{bbox}{bounding box of the mask, sets hard boundaries on the search
area of valid locations as well}

\item{sf}{return the land mask as an 'sf' polygon, not a rasterized map.
Mostly used in map plotting, not used for processing (default = FALSE)}
}
\value{
buffered land mask as an 'sf' or 'terra' map object
}
\description{
returns a buffered land mask to constrain potential
model results
}
\examples{
\dontrun{
# define land mask with a bounding box
# and an off-shore buffer (in km), in addition
# you can specifiy the resolution of the resulting raster
mask <- stk_mask(
  bbox  =  c(-20, -40, 60, 60), #xmin, ymin, xmax, ymax
  buffer = 150, # in km
  resolution = 0.5 # map grid in degrees
  )
}
}
