% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stk_mask.R
\name{stk_mask}
\alias{stk_mask}
\title{A land mask}
\usage{
stk_mask(buffer = 0)
}
\arguments{
\item{buffer}{buffer in degrees}
}
\value{
buffered land mask as an sf object
}
\description{
returns a buffered land mask to constrain potential
model results
}
