% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stk_profile.R
\name{stk_profile}
\alias{stk_profile}
\title{Plot seasonal profiles}
\usage{
stk_profile(data, logger, range = c(0, 1e+05), center = "day", plotly = FALSE)
}
\arguments{
\item{data}{skytrackr compatible data frame}

\item{logger}{the logger to plot}

\item{range}{the light range to plot}

\item{center}{center data on day or night}

\item{plotly}{TRUE or FALSE (convert to dynamic plotly plot or not)}
}
\value{
a plotly dynamic graph of light levels for a given logger
}
\description{
Uses plotly to provide a seasonal profile plot
NOTE make this a method class so you can call
plot(data.frame) or plot(df, plotly = TRUE)
}
