% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stk_read.R
\name{stk_read_lux}
\alias{stk_read_lux}
\title{Read Migrate Technology .lux files}
\usage{
stk_read_lux(files, verbose = TRUE)
}
\arguments{
\item{files}{A `.lux` file or list of `.lux` files with light level values}

\item{verbose}{provide detailed feedback}
}
\value{
A skytrackr compatible data frame for use in further location
 estimation.
}
\description{
Read Migrate Technology Ltd. `.lux` files and re-formats
them to a skytrackr compatible format.
}
\examples{
# read in the demo lux file
df <- stk_read_lux(
 system.file("extdata/cc876.lux", package="skytrackr")
)
}
